+++ crates/emscripten-stdio/src/detail.rs
